Document draw-column-headers
While doing so, I realized the way I was doing headers for big-endian
tables was way too esoteric for the kinds of beginners who use this
library, so I also changed the implementation to offer a much cleaner
way to accomplish that, and showed examples of how to use it. This
will require a new release of the node module.

Plus, the function didn't actually have a zero-arity version; it must
be a feature of the `sci` interpreter that was letting me get away
with calling it that way!